+++ CHANGELOG.md
@@ -12,7 +12,8 @@ section below it for the last release. -->
 * [GNATformat](https://github.com/AdaCore/gnatformat) is now the default back-end for LSP formatting requests
 * Add tasks and a CodeLens to run and debug a given main with GNATemulator on non-native projects
 * Diagnostics are now emitted for issues encountered when trying to load an Alire crate
-* Commands to create a new main units and packages have been added, both available under the `File->New File...` menu
+* Commands to create a new main units, packages and GPR project files have
+been added, both available under the `File->New File...` menu
 * Add VS Code command `ada: GNATcoverage - Load an existing XML coverage report` for importing coverage reports
 * Support running GNATtest tests in coverage mode